wvoigt language file updates for v2 text + fixed #72 (#106)
Added v2 text on the cards in brackets [ ] to indicate that this is newer text.  
Added fix for #72 
Fixed up printing problem with debt and coins on the tab.  Debt was being printed first instead of the coins as is on the cards.  Also changed the mask for debt to remove most near white background so it appears better on dark backgrounds.

* Update cards_fr.json

* Update cards_it.json

* Update cards_xx.json

* Add Changes to fix #72

* Update cards.py

* Update draw.py

Fixed issues with cost printing debt first, then coin (should be the other way around).  Also fixed extra so it will print \n now.

* Update __init__.py

* Fix PEP-8 issues

* Fix PEP-8

* PEP-8 fix