Fix Prosperity rules text running over
This mostly fixes Issue #155.  This seems to occur only with the VP image.  I've seen several coin and potion, and debt images at the end of a line and they are justified correctly.  There is some strange interaction with the VP image and the paragraph wrap function in report.
While not completely fixing the image out of the normal paragraph boundary, at least this keeps it within the margins of the card.