Fix for issue #239
Fix for issue #239.  If only upgrade expansions are selected, and the --upgrade-with-expansion is used, no cards were selected.  This happens because the upgrade cards are moved under the original expansion, when has not be selected for printing.  And since only the upgrade expansion was selected for printing, nothing now prints.
The fix changes this to now include the original expansion in this case.  So now this results in the original expansion expansion being printed with the upgrade expansion cards included as well.