Fix --expansion-reset-tabs limitations
The --expansion-reset-tabs option was previously incompatible with
--include-blanks because it handles low card counts specially, and
blanks leave the "extras" card count unset. I fixed the code by
defaulting to a count of 0 in this case.

The option was also previously dependent on --expansion-dividers, and
silently ignored if the latter was not set. As there wasn't a clear
technical reason for the limitation, I removed it and moved
--expansion-reset-tabs from the Expansion Dividers group to the Divider
Tab group.